Add (and use) ConfigureFunctionsFramework extension method
This allows for user customization of the host builder before we get to it (via a custom entry point)

Fixes #181
